+++ src/constants/RTE_MEMZONE.rs
@@ -2,13 +2,13 @@
 // Copyright © 2016 The developers of dpdk-sys. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/dpdk-sys/master/COPYRIGHT.
 
 
-pub const RTE_MEMZONE_16GB: c_int = 512;
-pub const RTE_MEMZONE_16MB: c_int = 256;
-pub const RTE_MEMZONE_1GB: c_int = 2;
-pub const RTE_MEMZONE_256KB: c_int = 65536;
-pub const RTE_MEMZONE_256MB: c_int = 131072;
-pub const RTE_MEMZONE_2MB: c_int = 1;
-pub const RTE_MEMZONE_4GB: c_int = 524288;
-pub const RTE_MEMZONE_512MB: c_int = 262144;
-pub const RTE_MEMZONE_NAMESIZE: c_int = 32;
-pub const RTE_MEMZONE_SIZE_HINT_ONLY: c_int = 4;
+pub const RTE_MEMZONE_16GB: c_uint = 512;
+pub const RTE_MEMZONE_16MB: c_uint = 256;
+pub const RTE_MEMZONE_1GB: c_uint = 2;
+pub const RTE_MEMZONE_256KB: c_uint = 65536;
+pub const RTE_MEMZONE_256MB: c_uint = 131072;
+pub const RTE_MEMZONE_2MB: c_uint = 1;
+pub const RTE_MEMZONE_4GB: c_uint = 524288;
+pub const RTE_MEMZONE_512MB: c_uint = 262144;
+pub const RTE_MEMZONE_NAMESIZE: size_t = 32;
+pub const RTE_MEMZONE_SIZE_HINT_ONLY: c_uint = 4;
